5 test cases passed , wrong files to be tested

+++ test/test5.c
@@ -1,4 +1,4 @@
-float add_float(float a1, float a2) //test float
+double add_float(double a1, double a2) //test double
 {
     return a1+a2;
 }
@@ -6,10 +6,10 @@ float add_float(float a1, float a2) //test float
 int main()
 {
     int i=0,t=2;
-    float f1=2.0;
-    float f2=3.0;
+    double f1=2.0;
+    double f2=3.0;
 
-    float re=add_float(f1,f2);
+    double re=add_float(f1,f2);
 
     do //test do-while
     {
